Add switch to display scores based on the Teacher/Student ResultState

+++ lib/models/rubric.dart
@@ -1,21 +1,21 @@
 class Rubric {
   String name;
   // Add other properties as needed
-  int score;
+  double score;
 
   Rubric({required this.name, required this.score});
 
   factory Rubric.fromJson(Map<String, dynamic> json) {
     return Rubric(
       name: json['name'],
-      score: json['score'],
+      score: json['score'].toDouble(),
     );
   }
 
   factory Rubric.fromMap(Map<String, dynamic> map) {
     return Rubric(
       name: map['name'],
-      score: map['score'],
+      score: map['score'].toDouble(),
     );
   }
